requirements.txt
git+https://github.com/Rapptz/discord.py#egg=discord.py==2.0.0a3575+g45d498c
PyYAML==6.0
pytz==2021.3
